fixed hit_circle (return value is double !\)& changed short if

+++ src/include/ray.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ray.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jhwang2 <jhwang2@student.42seoul.kr>       +#+  +:+       +#+        */
+/*   By: jeelee <jeelee@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/14 17:37:53 by jeelee            #+#    #+#             */
-/*   Updated: 2023/07/09 18:41:43 by jhwang2          ###   ########.fr       */
+/*   Updated: 2023/07/10 03:59:51 by jeelee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ int			hit_sphere(t_ray *ray, t_object *obj, double value[]);
 int			hit_plane(t_ray *ray, t_object *obj, double value[]);
 int			hit_cylinder(t_ray *ray, t_object *obj, double value[]);
 int			hit_cone(t_ray *ray, t_object *obj, double value[]);
-int			hit_circle(t_ray *ray, t_object *obj);
+double		hit_circle(t_ray *ray, t_object *obj);
 
 t_rec		get_intersection(t_ray *ray, t_object *obj);
 t_rec		find_closestobj(t_ray *ray, t_object **objs);
